+++ node/src/types/kernel/api/block/getDocInfo/index.d.ts
@@ -17,18 +17,7 @@
 
 
 /* directories */
-export * as appendBlock from "./appendBlock";
-export * as deleteBlock from "./deleteBlock";
-export * as getBlockBreadcrumb from "./getBlockBreadcrumb";
-export * as getBlockDOM from "./getBlockDOM";
-export * as getBlockInfo from "./getBlockInfo";
-export * as getBlockKramdown from "./getBlockKramdown";
-export * as getChildBlocks from "./getChildBlocks";
-export * as getDocInfo from "./getDocInfo";
-export * as insertBlock from "./insertBlock";
-export * as moveBlock from "./moveBlock";
-export * as prependBlock from "./prependBlock";
-export * as transferBlockRef from "./transferBlockRef";
-export * as updateBlock from "./updateBlock";
 
 /* files */
+export * from "./payload";
+export * from "./response";